Add parent_alive; default child sub uses it
The parent_alive checks if the daemon parent process is still alive,
which is important so children can commit suicide when they become
orphaned zombies. (Such a lovely thought. Great movie concept right
there.) Also, improved the tests so they are a bit more forgiving
if there's not exactly the same set of signal responses coming off
a daemon depending on exactly when it receives a shutdown command.